added functions for control of input id and new hamsterobject

+++ routes/hamsters.js
@@ -15,43 +15,62 @@ async function getHamsters() {
 	return hamsters;
 }
 
+async function checkId(inputId) {
+	const hamsters = await getHamsters();
+
+	return hamsters.find(hamster => hamster.firestoreId === inputId);
+}
+
+async function checkObject(inputObject) {
+	return Object.keys(inputObject).length !== 0;
+}
+
 router.get('/', async (req, res) => {
 	const hamsters = await getHamsters();
 
-	res.send(hamsters);
+	res.status(200).send(hamsters);
 });
 
 router.get('/random', async (req, res) => {
 	const hamsters = await getHamsters();
-
 	let randomIndex = Math.floor(Math.random() * (hamsters.length -1));
-	res.send(hamsters[randomIndex]);
+
+	res.status(200).send(hamsters[randomIndex]);
 });
 
 router.get('/:id', async (req, res) => {
 	const id = req.params.id;
-	const docRef = await db.collection('hamsters').doc(id).get();
+	const isCorrectId = await checkId(id);
 
+	if (!isCorrectId) {
+		res.status(400).send("Kontrollera ditt hamster id");
+		return;
+	}
+	const docRef = await db.collection('hamsters').doc(id).get();
 	const hamster = docRef.data();
 
-	res.send(hamster);
+	res.status(200).send(hamster);
 });
 
 router.post('/', async (req, res) => {
 	const object = req.body;
-
+	const isObject = await checkObject(object);
+	if (!isObject) {
+		res.status(400).send("Kontrollera hamsterobjektet du försöker lägga till");
+		return;
+	}
 	const docRef = await db.collection('hamsters').add(object);
-	res.send(docRef.id);
+
+	res.status(200).send(`Du har lagt till en ny hamster med id: ${docRef.id}`);
 });
 
 router.post('/postmany', async (req, res) => {
 	const array = req.body;
-	
 	await array.forEach(hamster => {
 	db.collection('hamsters').add(hamster);
 	});
 
-	res.send("Dina hamstrar är tillagda");
+	res.status(200).send("Dina hamstrar är tillagda");
 });
 
 router.put('/', (req, res) => {
@@ -61,9 +80,14 @@ router.put('/', (req, res) => {
 router.put('/:id', async (req, res) => {
 	const object = req.body;
 	const id = req.params.id;
-
+	const isCorrectId = await checkId(id);
+	if (!isCorrectId) {
+		res.status(400).send("Kontrollera ditt hamster id");
+		return;
+	}
 	const docRef = db.collection('hamsters').doc(id);
 	await docRef.set(object, {merge: true});
+
 	res.status(200).send("Hamstern är ändrad");
 });
 
@@ -73,8 +97,13 @@ router.delete('/', (req, res) => {
 
 router.delete('/:id', async (req, res) => {
 	const id = req.params.id;
-
+	const isCorrectId = await checkId(id);
+	if (!isCorrectId) {
+		res.status(400).send("Kontrollera ditt hamster id");
+		return;
+	}
 	await db.collection('hamsters').doc(id).delete();
+
 	res.status(200).send("Hamstern är borttagen");
 });
